+++ mountpoint-s3-fs/src/logging.rs
@@ -6,17 +6,28 @@ use std::panic::{self, PanicHookInfo};
 use std::path::{Path, PathBuf};
 use std::thread;
 
-use crate::metrics::metrics_tracing_span_layer;
 use anyhow::Context;
-use mountpoint_s3_client::config::RustLogAdapter;
 use rand::Rng;
+use signal_hook::consts::SIGUSR2;
+use signal_hook::iterator::{Handle as SignalsHandle, Signals};
 use time::format_description::FormatItem;
 use time::macros;
 use time::OffsetDateTime;
 use tracing::Span;
+use tracing_log::log::warn;
 use tracing_subscriber::filter::{EnvFilter, LevelFilter};
 use tracing_subscriber::layer::SubscriberExt;
 use tracing_subscriber::util::SubscriberInitExt;
+
+use mountpoint_s3_client::config::{RustLogAdapter, AWSCRT_LOG_TARGET};
+
+use crate::metrics::metrics_tracing_span_layer;
+
+#[cfg(test)]
+mod testing;
+
+mod envfilter;
+use envfilter::{toggleable, ToggleableHandle};
 
 mod syslog;
 use self::syslog::SyslogLayer;
@@ -35,16 +46,28 @@ pub struct LoggingConfig {
     pub default_filter: String,
 }
 
+#[derive(Default)]
+/// A handle for logging that cleans up all allocated resources on drop.
+pub struct LoggingHandle {
+    _toggle_signal_handle: Option<ToggleSignalHandle>,
+}
+
 /// Set up all our logging infrastructure.
 ///
 /// This method:
 /// - initializes the `tracing` subscriber for capturing log output
 /// - sets up the logging adapters for the CRT and for metrics
 /// - installs a panic hook to capture panics and log them with `tracing`
-pub fn init_logging(config: LoggingConfig) -> anyhow::Result<()> {
-    init_tracing_subscriber(config)?;
+/// - sets up a signal listener and toggles the logging verbosity each time it receives a `USR2` signal
+pub fn init_logging(config: LoggingConfig) -> anyhow::Result<LoggingHandle> {
+    let handle = init_tracing_subscriber(config)?;
     install_panic_hook();
-    Ok(())
+    Ok(handle)
+}
+
+/// Record the object name in the current [Span].
+pub fn record_name(name: &str) {
+    Span::current().record("name", name);
 }
 
 /// For a given log directory, prepare a file name for this Mountpoint.
@@ -100,14 +123,11 @@ fn install_panic_hook() {
     }))
 }
 
-fn init_tracing_subscriber(config: LoggingConfig) -> anyhow::Result<()> {
-    // Create the logging config from the MOUNTPOINT_LOG environment variable or the default config
-    // if that variable is unset.
-    let env_filter =
-        EnvFilter::try_from_env("MOUNTPOINT_LOG").unwrap_or_else(|_| EnvFilter::new(&config.default_filter));
+fn init_tracing_subscriber(config: LoggingConfig) -> anyhow::Result<LoggingHandle> {
     // Don't create the files or subscribers if we'll never emit any logs
-    if env_filter.max_level_hint() == Some(LevelFilter::OFF) {
-        return Ok(());
+    let default_filter = make_default_filter(config.default_filter.clone())();
+    if default_filter.max_level_hint() == Some(LevelFilter::OFF) {
+        return Ok(LoggingHandle::default());
     }
 
     RustLogAdapter::try_init().context("failed to initialize CRT logger")?;
@@ -143,17 +163,91 @@ fn init_tracing_subscriber(config: LoggingConfig) -> anyhow::Result<()> {
         None
     };
 
+    let (filter, filter_handle) = toggleable(vec![
+        // Default logging verbosity (i.e., the one configured using `--debug`, `--debug-crt`, or `MOUNTPOINT_LOG` environment variable)
+        make_default_filter(config.default_filter),
+        // Debug logging for all except CRT (i.e., `debug,awscrt=off`)
+        make_filter(LevelFilter::DEBUG, LevelFilter::OFF),
+        // Debug logging for all (i.e., `debug,awscrt=debug`)
+        make_filter(LevelFilter::DEBUG, LevelFilter::DEBUG),
+        // Trace logging for all except CRT (i.e., `trace,awscrt=off`)
+        make_filter(LevelFilter::TRACE, LevelFilter::OFF),
+        // Trace logging for all (i.e., `trace,awscrt=trace`)
+        make_filter(LevelFilter::TRACE, LevelFilter::TRACE),
+    ]);
+
+    let toggle_signal_handle = toggle_filter_on_signals(vec![SIGUSR2], filter_handle)?;
+
     tracing_subscriber::registry()
-        .with(env_filter)
+        .with(filter)
         .with(syslog_layer)
         .with(file_layer)
         .with(console_layer)
         .with(metrics_tracing_span_layer())
         .init();
 
-    Ok(())
+    Ok(LoggingHandle {
+        _toggle_signal_handle: Some(toggle_signal_handle),
+    })
 }
 
-pub fn record_name(name: &str) {
-    Span::current().record("name", name);
+/// Create a default logging filter from the `MOUNTPOINT_LOG` environment variable or the default config
+/// if that variable is unset.
+fn make_default_filter(default_directives: String) -> Box<dyn FnMut() -> EnvFilter + Send> {
+    Box::new(move || {
+        EnvFilter::try_from_env("MOUNTPOINT_LOG").unwrap_or_else(|_| EnvFilter::new(default_directives.clone()))
+    })
+}
+
+/// Create a logging filter using provided global and CRT [level](LevelFilter)s.
+fn make_filter(level: LevelFilter, crt_level: LevelFilter) -> Box<dyn FnMut() -> EnvFilter + Send> {
+    Box::new(move || EnvFilter::new(format!("{level},{AWSCRT_LOG_TARGET}={crt_level}")))
+}
+
+/// Creates a new thread to listen specified Unix signals, and toggles the log filter each time it receives a signal.
+/// Returns a [handle](ToggleSignalHandle), and cleans up signal listener and the created thread once the handle is dropped.
+fn toggle_filter_on_signals<S: 'static>(
+    signals: Vec<Signal>,
+    mut toggle_handle: ToggleableHandle<S>,
+) -> anyhow::Result<ToggleSignalHandle> {
+    let mut signals = Signals::new(signals)?;
+    let signals_handle = signals.handle();
+
+    let thread_handle = thread::spawn(move || {
+        for _ in &mut signals.forever() {
+            match toggle_handle.next() {
+                Ok(desc) => {
+                    warn!("Changed log verbosity to {desc}");
+                }
+                Err(err) => {
+                    warn!("Failed to change log verbosity: {err}");
+                }
+            }
+        }
+    });
+
+    Ok(ToggleSignalHandle {
+        signals_handle,
+        thread_handle: Some(thread_handle),
+    })
+}
+
+/// An Unix signal.
+type Signal = libc::c_int;
+
+/// A handle returned by [toggle_filter_on_signals] which cleans up all allocated resources on [drop](ToggleSignalHandle::drop).
+struct ToggleSignalHandle {
+    signals_handle: SignalsHandle,
+    thread_handle: Option<thread::JoinHandle<()>>,
+}
+
+impl Drop for ToggleSignalHandle {
+    fn drop(&mut self) {
+        if !self.signals_handle.is_closed() {
+            self.signals_handle.close();
+        }
+        if let Some(handle) = self.thread_handle.take() {
+            _ = handle.join();
+        }
+    }
 }